refactor(martinez): position polygon-Boolean engine as INTERNAL infrastructure
This commit reframes the Martinez clipping work as in-process infrastructure
for the temporal types whose values are 2D regions (trgeo, tcbuffer,
tgeometry, tgeography), rather than as a public ST_*-competing SQL surface.
This is a strategic positioning change driven by two observations:

1. The fast trajectory-clip work (tgeo_fast_clip.c, on origin/tcbuffer_fix)
   only handles tgeompoint/tgeogpoint — temporal values that are points
   moving along line segments. For trgeo/tcbuffer/tgeometry/tgeography the
   temporal value at each instant is itself a 2D region (rigid polygon,
   circle, or arbitrary geometry); their `atGeometry` reduces to a sequence
   of polygon-vs-polygon Boolean ops per segment. THAT is the workload that
   needs an in-process engine — the GEOS GSERIALIZED <-> GEOS roundtrip
   would be paid millions of times in real queries (e.g. 10K-row trgeo
   table * 100 instants/row = 1M Boolean ops per atGeometry call).

2. As a public ST_* replacement the four cl_* functions add no unique
   capability: PostGIS ST_Intersection / ST_Union / ST_Difference /
   ST_SymDifference are more robust on degenerate inputs and handle curves,
   geography and 3D. The audit-flagged regressions (deleted-flag splay-tree
   bug at L927-958, connect_edges premature termination at L1276-1295,
   self-intersection unhandled, EPS-relative gates omitted) would surface
   in user workloads, and the maintenance team would own them. The
   GEOS-roundtrip cost argument that justifies tgeo_fast_clip does not
   transfer to one-shot user polygon-Boolean queries — it amortizes.

Changes:

- File rename: meos/{src,include}/geo/tpoint_clipping.{c,h} ->
  meos/{src,include}/geo/geo_poly_clip.{c,h}. The "tpoint" prefix was
  misleading: there is no tpoint code in the file, just polygon Boolean
  ops on GSERIALIZED. The new name reflects that this is a geometry
  primitive intended for internal MEOS use.

- SQL surface demotion: rename the four CREATE FUNCTION declarations from
  cl_intersection / cl_union / cl_difference / cl_symDifference to
  _mdb_internal_clip_intersection / _union / _difference /
  _sym_difference. The leading underscore + _mdb_internal_ prefix follows
  the convention for "implementation detail, not part of the public API".
  Header comment explicitly directs end users to PostGIS ST_*.
  C function symbols (cl_intersection etc.) are unchanged so the existing
  PG_FUNCTION_INFO_V1 wrappers in tgeo_spatialfuncs.c keep working.

- Test file updated to use the new internal SQL names.

Build: cmake --build clean, zero warnings; full extension links; both
the renamed C-engine and the four renamed-SQL wrappers exported.

NOT changed (next commits):
- The audit-flagged structural bugs in the algorithm itself. These were
  P2 ("defer until we have a reproducer") when no caller existed; once
  trgeo/tcbuffer/tgeometry route through this engine they become P0,
  but they need a known-bad-input reproducer to validate any fix and
  risk new regressions if touched blind.
- Wiring the C engine into trgeo/tcbuffer/tgeometry restriction code
  paths. That requires identifying the specific GEOS call sites in those
  modules (lwgeom_intersection / lwgeom_union calls) and is a separate
  unit of work.
