+++ mobilitydb/sql/geo/056_tpoint_spatialfuncs.in.sql
@@ -352,21 +352,29 @@ CREATE FUNCTION minusStbox(tgeogpoint, stbox, borderInc bool DEFAULT TRUE)
 
 /*****************************************************************************
  * Polygon clipping (Martinez-Rueda algorithm)
+ *
+ * INTERNAL — these functions expose the in-process polygon Boolean engine
+ * used by the temporal types whose values are 2D regions (trgeo, tcbuffer,
+ * tgeometry, tgeography). They exist so the engine can be unit-tested and
+ * benchmarked from SQL; end users should keep using PostGIS ST_Intersection
+ * / ST_Union / ST_Difference / ST_SymDifference, which are more robust on
+ * degenerate inputs and handle curves / geography / 3D. The `_mdb_internal_`
+ * prefix marks them as not part of the public API.
  *****************************************************************************/
 
-CREATE FUNCTION cl_intersection(geometry, geometry)
+CREATE FUNCTION _mdb_internal_clip_intersection(geometry, geometry)
   RETURNS geometry
   AS 'MODULE_PATHNAME', 'cl_intersection'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
-CREATE FUNCTION cl_union(geometry, geometry)
+CREATE FUNCTION _mdb_internal_clip_union(geometry, geometry)
   RETURNS geometry
   AS 'MODULE_PATHNAME', 'cl_union'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
-CREATE FUNCTION cl_difference(geometry, geometry)
+CREATE FUNCTION _mdb_internal_clip_difference(geometry, geometry)
   RETURNS geometry
   AS 'MODULE_PATHNAME', 'cl_difference'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
-CREATE FUNCTION cl_symDifference(geometry, geometry)
+CREATE FUNCTION _mdb_internal_clip_sym_difference(geometry, geometry)
   RETURNS geometry
   AS 'MODULE_PATHNAME', 'cl_symDifference'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
