+++ mobilitydb/test/geo/queries/078_tpoint_clipping.test.sql
@@ -32,37 +32,37 @@
 -------------------------------------------------------------------------------
 
 -- Point intersections
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
   geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))'));
 
 -- Overlapping Segments
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
   geometry 'Polygon((3 1,3 5,7 5,7 1,3 1))'));
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
   geometry 'Polygon((1 3,1 7,5 7,5 3,1 3))'));
-SELECT st_astext(cl_intersection(geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))',
   geometry 'Polygon((1 4,4 1,7 4,4 7,1 4))'));
 
 -- Point and segment intersections
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
   geometry 'Polygon((3 1,3 6,7 6,7 1,3 1))'));
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1))',
   geometry 'Polygon((1 3,1 7,6 8,6 3,1 3))'));
-SELECT st_astext(cl_intersection(geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))',
   geometry 'Polygon((0 5,4 1,7 4,3 8,0 5))'));
 
 -- Hole that does not interact with the other polygon
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,3 4,3 2,2 2))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,3 4,3 2,2 2))',
   geometry 'Polygon((4 1,4 5,8 5,8 1,4 1))'));
 -- Hole whose countour intersects with the contour of the other polygon
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,3 4,3 2,2 2))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,3 4,3 2,2 2))',
   geometry 'Polygon((3 1,3 5,6 5,6 1,3 1))'));
 -- Hole that removes part of the contour of the other polygon
-SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,4 4,4 2,2 2))',
+SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,2 4,4 4,4 2,2 2))',
   geometry 'Polygon((3 1,3 5,6 5,6 1,3 1))'));
 
 -- Point intersections with a hole that does not interact with the other polygon
--- SELECT st_astext(cl_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,4 2,4 4,2 4,2 2))',
+-- SELECT st_astext(_mdb_internal_clip_intersection(geometry 'Polygon((1 1,1 5,5 5,5 1,1 1),(2 2,4 2,4 4,2 4,2 2))',
   -- geometry 'Polygon((0 3,3 0,6 3,3 6,0 3))'));
 
 --------------------------------------------------------
